pigfarm_chatbot/requirements.txt
# FastAPI & Server
fastapi[standard]>=0.115.0
sse-starlette>=2.1.0

# LangChain (Latest)
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
pgvector>=0.3.0

# Document Processing
pypdf>=4.3.0
tiktoken>=0.7.0

# Reranking
cohere>=5.9.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.8.0
pydantic-settings>=2.4.0
numpy>=1.26.4
google-generativeai>=0.8.0

# Observability & Evaluation
langsmith>=0.1.17
